backend/requirements.txt
Django>=6.0.0
djangorestframework>=3.17.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.9.0
psycopg2-binary>=2.9.12
python-dotenv>=1.2.0
Pillow>=12.0.0
django-filter>=25.2
requests>=2.33.0
